Refactor, again
Dictionary full of nonsense was unmaintainable, write a class for each mod record.

Get the mod list out of the main loop, into the process loop

Add column sorting (yah!)

Add size reports.  Full path is stored too, maybe we can do something with this later.

Write a wrapper around the log.  That might need to be a class too.  We'll see. (for reference, it's just and ordered list that is built the same time the UI is, then dumped to a text file on demand)