Rework conflicts, add "unzip" style bad mod check
For mods that have "unzip" in the name, chances are you should follow instructions.  Warn on these

Rework the conflict system to accept a second list of mods that the listed mod conflicts with.  Make the data storage on this a dictionary of dictionaries instead of a simple dict.  Closes #10