Add file test - all
Test all "mods" to see if they are valid.  We already were for folders, might as well do the lot.  On *massive* collections, this is gonna slow things down.  Probably a lot.  My collection of 150 / 6Gb, it's about a second, maybe 1.5 seconds.

Don't track size of bad mods, we don't care.