Bad File check - add sha sum
When checking a poorly named file, do the following:

1.) Try and guess the name, then:

2.) If the guess is a real, found mod, then:

3.) Compare the version strings, then:

4.) If those are the same, compare SHA256 sums to make sure the files are identical

Note: we could reverse the order of 3 and 4, but at that point the version is already loaded (or it has failed to load), so that comparison costs us nothing. If it failed to find the version, it already errored out earlier as not valid.  And if the versions are different, sha256 will fail anyway. Getting a sum isn't particularly IO expensive, but we still don't want to have to do it $number_of_mods if not necessary.