Catch mods that are both a file and a zip folder
Because we want to do what the game does, we actually only test the zip. Due to order of processing, it still gets added to the "bad" list and the user gets warned to zip the folder.  We also set a special flag so we can add this to the "conflicts" tab and warn that the game is going to load the zip rather than the (possibly changed) folder.

Closes #19