Fix edge case: copy exists, original missing but in savegame
Created a condition where the original file "exists", but cannot be opened (since it's not really there).  Caused the MD5 hashing to fail.

Now, first check is if the original is referenced, then check if it's a real file, finally get the sum.

Also harden up the MD5 sum section to soft-fail with logging