Fix for wrong language selected...
When the user's computer is set to a language mod check doesn't yet know, we fall back to english - but english is not selected in the language picker UI - due to how this list is built, and because we want to continue to fall back to english, not explicitly set the language to english, I re-ordered the language list picker so english is first, fixing this visual problem.  All others are in alphabetical order as before.